Add missing parenthesis to next() variable assignment in smv code
The right-hand side of assignments to the next state of an smv variable
must be parenthesized to avoid operator precedence issues.

For example, an assignment like [equal := a == b] where equal is a
boolean and a and b are integers, will cause a typing error in nuxmv due
to assigning an integer value to a boolean variable, due to equality
being left-associative.