Interpreter.ofTrace(): Check transitions
In the case of supply/get, the NuXmv trace can mention the labels of
the interacting agents and those of their respective actions
(e.g. agent1-mySupply-agent2-myGet)
When loading a trace we ensure that we pick that transition and not
merely the first one which is compatible with the resulting state.

In order to do that we need the *full* state at a given step, so we add
a class CompositeJSONObject that lazily merges a base JSON with a diff
(actually it acts as a sort of proxy towards these two object, so no
data is copied)