Specialize nuxmv model to individual LTOL
We now create a separate nuxmv model file for each LTOL formula in the
specifications (it's cheap, anyway), so that we only add observation
variables to the model if that specific formula requires them.

Besides making some verification tasks a little lighter, this also
fixes some bug when loading counterexamples into the interpreter.